# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mencari apotek terdekat, reservasi obat, dan transaksi pembayaran",
        "title": "Medifinder API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "cs.medifinder@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/obat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin Obat"
                ],
                "summary": "Lihat stok obat saya (Admin Only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin apotek menambahkan obat ke apotek miliknya",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Obat"
                ],
                "summary": "Tambahkan obat baru",
                "parameters": [
                    {
                        "description": "Data obat",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ObatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/obat/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update informasi nama, stok, atau harga obat oleh admin apotek",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Obat"
                ],
                "summary": "Update data obat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Obat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update obat",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ObatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: obat updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: invalid input / error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data obat milik admin apotek",
                "tags": [
                    "Admin Obat"
                ],
                "summary": "Hapus obat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Obat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "obat deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/apotek/{id}/obat": {
            "get": {
                "tags": [
                    "Obat"
                ],
                "summary": "Lihat daftar obat di apotek tertentu (Public)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Apotek ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search by Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua daftar obat yang ada di keranjang user saat ini",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keranjang"
                ],
                "summary": "Lihat isi keranjang",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CartResponse"
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cart/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat transaksi baru dari item dalam keranjang dan mengembalikan token pembayaran",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keranjang"
                ],
                "summary": "Checkout keranjang",
                "responses": {
                    "200": {
                        "description": "transaksi_id, snap_token, redirect_url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "cart kosong / stok tidak cukup",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cart/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keranjang"
                ],
                "summary": "Tambah obat ke keranjang",
                "parameters": [
                    {
                        "description": "Data item keranjang",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cart/items/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keranjang"
                ],
                "summary": "Update jumlah item di keranjang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cart Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update quantity",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus satu item obat dari keranjang belanja",
                "tags": [
                    "Keranjang"
                ],
                "summary": "Hapus item dari keranjang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cart Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: item deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.AddCartRequest": {
            "type": "object",
            "properties": {
                "jumlah": {
                    "type": "integer",
                    "example": 2
                },
                "obat_id": {
                    "type": "string",
                    "example": "0d744571-..."
                }
            }
        },
...
        "dto.CartItemResponse": {
            "type": "object",
            "properties": {
                "harga": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "jumlah": {
                    "type": "integer"
                },
                "nama_obat": {
                    "type": "string"
                },
                "obat_id": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "dto.CartResponse": {
            "type": "object",
            "properties": {
                "apotek_id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CartItemResponse"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.ObatRequest": {
            "type": "object",
            "properties": {
                "harga": {
                    "type": "number",
                    "example": 5000
                },
                "nama": {
                    "type": "string",
                    "example": "Paracetamol"
                },
                "stok": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "dto.UpdateCartRequest": {
            "type": "object",
            "properties": {
                "jumlah": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}